Moving classes to crud
git-svn-id: https://opendentalsoft.com:23793/svn/opendental/head@3936 098d0599-7f2b-0410-95a5-ad30ac85393a

+++ OpenDental/Main Modules/ContrAccount.cs
@@ -2765,7 +2765,7 @@ namespace OpenDental {
 						claimProcs[i].CodeSent=claimProcs[i].CodeSent.Substring(0,5);
 					}
 				}
-				claimProcs[i].LineNumber=i+1;
+				claimProcs[i].LineNumber=(byte)(i+1);
 				ClaimProcs.Update(claimProcs[i]);
 			}//for claimProc
 			return ClaimCur;
